--- a/R/rng.R
+++ b/R/rng.R
@@ -181,7 +181,7 @@ recycle <- function(x, n, name = deparse(substitute(x))) {
   if (length(x) == n) {
     x
   } else if (length(x) == 1L) {
-    rep(x, n)
+    rep_len(x, n)
   } else {
     stop(sprintf("Invalid length for '%s', expected 1 or %d", name, n))
   }