--- a/inst/include/dust/distr/binomial.hpp
+++ b/inst/include/dust/distr/binomial.hpp
@@ -6,22 +6,50 @@
 namespace dust {
 namespace distr {
 
+// Faster version of pow(x, n) for integer 'n' by using
+// "exponentiation by squaring"
+// https://en.wikipedia.org/wiki/Exponentiation_by_squaring
 template <typename T>
-double binomial_inversion(rng_state_t<T>& rng_state,
-                          double n, double prob) {
-  double geom_sum = 0;
-  double num_geom = 0;
+T fast_pow(T x, int n) {
+  T pow = 1.0;
+  if (n != 0) {
+    for(;;) {
+      if(n & 01) {
+        pow *= x;
+      }
+      if(n >>= 1) {
+        x *= x;
+      } else {
+        break;
+      }
+    }
+  }
+  return pow;
+}
 
-  while (true) {
-    double r = dust::unif_rand<T, double>(rng_state);
-    double geom = std::ceil(std::log(r) / std::log1p(-prob));
-    geom_sum += geom;
-    if (geom_sum > n) {
+// Binomial random numbers via inversion (for low np only!). Draw a
+// random number from U(0, 1) and find the 'n' up the distribution
+// (given p) that corresponds to this
+template <typename T>
+T binomial_inversion(rng_state_t<T>& rng_state, int n, T p) {
+  T u = dust::unif_rand<T>(rng_state);
+
+  // This is equivalent to qbinom(u, n, p)
+  const T q = 1. - p;
+  const T r = p / q;
+  const T g = r * (n + 1);
+  int ix = 0;
+  T f = fast_pow(q, n);
+  for (;;) {
+    if (u < f) {
       break;
     }
-    ++num_geom;
+    u -= f;
+    ix++;
+    f *= (g / ix - r);
   }
-  return num_geom;
+
+  return ix;
 }
 
 inline double stirling_approx_tail(double k) {