--- a/inst/include/dust/distr/binomial.hpp
+++ b/inst/include/dust/distr/binomial.hpp
@@ -35,21 +35,21 @@ T binomial_inversion(rng_state_t<T>& rng_state, int n, T p) {
   T u = dust::unif_rand<T>(rng_state);
 
   // This is equivalent to qbinom(u, n, p)
-  const T q = 1. - p;
+  const T q = 1 - p;
   const T r = p / q;
   const T g = r * (n + 1);
-  int ix = 0;
-  T f = fast_pow(q, n);
+  const T f = fast_pow(q, n);
+  int k = 0;
   for (;;) {
     if (u < f) {
       break;
     }
     u -= f;
-    ix++;
-    f *= (g / ix - r);
+    k++;
+    f *= (g / k - r);
   }
 
-  return ix;
+  return k;
 }
 
 inline double stirling_approx_tail(double k) {