… R/dust.R
… src/sir.cpp
… src/variable.cpp
… src/volatility.cpp
… src/walk.cpp