This PR uses a faster algorithm for the binomial random number generation in the case of small np, primarily by using an exponentiation-by-squaring algorithm. I've also moved the `dust_rng.cpp` file over to use the cpp11 API directly which reduces a number of copies and brings our examples under the R timings in all cases on my laptop.

~Depends on #90~
Fixes #91