@richfitz 2020-08-17
A preliminary check with sircovid2 suggests that this might be worth ~10% performance

old:

```
# A tibble: 1 x 13
  expression        min  median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result  memory      time    gc      
  <bch:expr>     <bch:> <bch:t>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>  <list>      <list>  <list>  
1 filter$run(pp)  4.21s   4.72s     0.217    7.57MB   0.0435    10     2        46s <dbl [… <Rprofmem[… <bch:t… <tibble…
```

new:

```
# A tibble: 1 x 13
  expression        min  median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result  memory      time    gc      
  <bch:expr>     <bch:> <bch:t>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>  <list>      <list>  <list>  
1 filter$run(pp)  4.08s   4.22s     0.238    7.67MB   0.0952    10     4        42s <dbl [… <Rprofmem[… <bch:t… <tibble…
```

(this is on the refactor of the ons paper task)